Move AppStructureRepeater into receiverwrapper
