+++ corehq/apps/domain/views.py
@@ -26,8 +26,7 @@ from dimagi.utils.django.email import send_HTML_email
 from dimagi.utils.web import get_ip, json_response
 from corehq.apps.users.views import require_can_edit_web_users
 from corehq.apps.receiverwrapper.forms import FormRepeaterForm
-from corehq.apps.receiverwrapper.models import FormRepeater, CaseRepeater, ShortFormRepeater
-from corehq.apps.app_manager.models import AppStructureRepeater
+from corehq.apps.receiverwrapper.models import FormRepeater, CaseRepeater, ShortFormRepeater, AppStructureRepeater
 from django.contrib import messages
 from django.views.decorators.http import require_POST
 import json
